[TASK] Make AnonymizationService independent from TS config
and load the config in the CommandController.
This allows custom usage of the service and decouples it from configuration.
